Bundle Preload files into a Module (PR #2396)
The files for Preload are distributed through the plugin. This PR bundles them together by:

- Pulling all of them into one module, i.e. inc/Engine/Preload
- Changes the namespaces to match
- Renames the class files to PSR-4, i.e. to match the class names
