+++ inc/classes/subscriber/Optimization/class-critical-css-subscriber.php
@@ -4,6 +4,9 @@ namespace WP_Rocket\Subscriber\Optimization;
 use WP_Rocket\Event_Management\Subscriber_Interface;
 use WP_Rocket\Optimization\CSS\Critical_CSS;
 use WP_Rocket\Admin\Options_Data;
+use FilesystemIterator;
+use UnexpectedValueException;
+
 
 defined( 'ABSPATH' ) || exit;
 
@@ -34,9 +37,11 @@ class Critical_CSS_Subscriber implements Subscriber_Interface {
 	 * @return array
 	 */
 	public static function get_subscribed_events() {
+		$slug = rocket_get_constant( 'WP_ROCKET_SLUG' );
+
 		return [
 			'admin_post_rocket_generate_critical_css' => 'init_critical_css_generation',
-			'update_option_' . WP_ROCKET_SLUG         => [
+			'update_option_' . $slug                  => [
 				[ 'generate_critical_css_on_activation', 11, 2 ],
 				[ 'stop_process_on_deactivation', 11, 2 ],
 			],
@@ -133,17 +138,29 @@ class Critical_CSS_Subscriber implements Subscriber_Interface {
 	 * @param array $value     New values for WP Rocket settings.
 	 */
 	public function generate_critical_css_on_activation( $old_value, $value ) {
-		if ( isset( $old_value['async_css'], $value['async_css'] ) && ( $old_value['async_css'] !== $value['async_css'] ) && 1 === (int) $value['async_css'] ) {
-			try {
-				if ( ( new \FilesystemIterator( $this->critical_css->get_critical_css_path(), \FilesystemIterator::SKIP_DOTS ) )->valid() ) {
-					return;
-				}
-			} catch ( \UnexpectedValueException $e ) {
+		if ( ! isset( $old_value['async_css'], $value['async_css'] ) || ( $old_value['async_css'] === $value['async_css'] ) || 1 !== (int) $value['async_css'] ) {
+			return;
+		}
+
+		$critical_css_path = $this->critical_css->get_critical_css_path();
+
+		// Check if the CPCSS path exists and create it.
+		if ( ! rocket_direct_filesystem()->is_dir( $critical_css_path ) ) {
+			rocket_mkdir_p( $critical_css_path );
+		}
+
+		try {
+			if ( ( new FilesystemIterator( $critical_css_path, FilesystemIterator::SKIP_DOTS ) )->valid() ) {
+				// Bail out if the folder is not empty.
 				return;
 			}
-
-			$this->critical_css->process_handler();
+		} catch ( UnexpectedValueException $e ) {
+			// Bail out when folder is invalid.
+			return;
 		}
+
+		// Generate the CPCSS files.
+		$this->critical_css->process_handler();
 	}
 
 	/**
