--- tests/Fixtures/Preload/Process_Wrapper.php
@@ -1,29 +1,0 @@
-<?php
-// phpcs:ignoreFile
-namespace WP_Rocket\Tests\Integration\Fixtures\Preload;
-
-use WP_Rocket\Preload\Full_Process;
-
-/**
- * Wrapper class used to test the results.
- */
-class Process_Wrapper extends Full_Process {
-	private $generatedKey;
-
-	public function save() {
-		$key = $this->generate_key();
-
-		if ( ! empty( $this->data ) ) {
-			update_site_option( $key, $this->data );
-			$this->generatedKey = $key;
-		} else {
-			$this->generatedKey = null;
-		}
-
-		return $this;
-	}
-
-	public function getGeneratedKey() {
-		return $this->generatedKey;
-	}
-}
